# docs/openapi-snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "jekko-server",
    "description": "",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/workspace": {
      "get": {
        "tags": [
          "crate::routes::workspace"
        ],
        "summary": "`GET /api/v1/workspace`.",
        "operationId": "list",
        "responses": {
          "200": {
            "description": "Workspaces"
          }
        }
      },
      "post": {
        "tags": [
          "crate::routes::workspace"
        ],
        "summary": "`POST /api/v1/workspace`.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceDescriptor"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created"
          }
        }
      }
    },
    "/api/v1/workspace/{id}": {
      "get": {
        "tags": [
          "crate::routes::workspace"
        ],
        "summary": "`GET /api/v1/workspace/:id`.",
        "operationId": "get_workspace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace"
          }
        }
      },
      "delete": {
        "tags": [
          "crate::routes::workspace"
        ],
        "summary": "`DELETE /api/v1/workspace/:id`.",
        "operationId": "remove",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Removed"
          }
        }
      }
    },
    "/api/v1/ws/{session_id}": {
      "get": {
        "tags": [
          "crate::routes::ws"
        ],
        "summary": "`GET /api/v1/ws/:session_id` (upgrade).",
        "operationId": "handler",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Upgraded"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "WorkspaceDescriptor": {
        "type": "object",
        "description": "Workspace descriptor.",
        "required": [
          "id",
          "name",
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "description": "Workspace metadata."
          },
          "id": {
            "type": "string",
            "description": "Workspace id."
          },
          "name": {
            "type": "string",
            "description": "Display name."
          }
        }
      }
    }
  }
}